requirements.txt
brotli==1.0.7
django-admin-sortable2==0.7
django-environ==0.4.5
django==2.1.8
dj-database-url==0.5.0
gevent==1.4.0
gunicorn==19.9.0
psycogreen==1.0.1
psycopg2==2.8.1
pytz==2019.1
whitenoise[brotli]==4.1.2
